refactor: rename weights->default_weights
